moved examine to separate js file
