+++ client/src/components/Console/InputPanel.js
@@ -17,7 +17,7 @@ import { useAuth0 } from "@auth0/auth0-react";
 import DiscoverableCalls, { callFunctionMap } from "../../clientUtilities/discoverablesCalls";
 import DiscoverableFunctions from "../../clientUtilities/discoverablesFunctions";
 import { lookAbout } from './js/look';
-
+import runExamine from './js/examine';
 
 //set up index for current position in userCommandsHistory
 let inputHistoryIndex;
@@ -212,52 +212,10 @@ function InputPanel({
                 giveItem(socket, item, target, user, location);
 
             } else if (findIn(input, actionCalls.examine)) {
-                const command = getOneOfTheseOffThat(actionCalls.examine, input.toLowerCase());
                 let toExamine = takeTheseOffThat(actionCalls.examine, input.toLowerCase());
+                const command = getOneOfTheseOffThat(actionCalls.examine, input.toLowerCase());
                 toExamine = takeTheseOffThat(['the', 'a', 'an'], toExamine)
-                console.log("You are attempting to examine", toExamine)
-
-                if (location.current.discoverables && toExamine.trim() !== '') {
-                    let discoverables = location.current.discoverables;
-                    let description;
-                    let exampleCommand;
-                    discoverables.forEach(discoverable => {
-                        discoverable.names.forEach(name => {
-                            if (name.startsWith(toExamine.toLowerCase()) && toExamine.trim() !== '') {
-                                console.log("You found the", name);
-                                description = discoverable.description;
-                                exampleCommand = discoverable.exampleCommand;
-                            }
-                        })
-                    })
-                    if (description) {
-                        setChatHistory(prevState => {
-                            if (exampleCommand) {
-                                return [...prevState,
-                                { type: 'displayed-stat', text: `You see ${description}` },
-                                { type: 'displayed-commands', text: `Try entering: ${exampleCommand}` }]
-                            } else {
-                                return [...prevState, { type: 'displayed-stat', text: `You see ${description}` }]
-                            }
-                        })
-                    } else {
-                        setChatHistory(prevState => { return [...prevState, { type: "displayed-error", text: "There's nothing to discover by that name" }] })
-                    }
-                } else if (toExamine.trim() === '') {
-                    setChatHistory(prevState => { return [...prevState, { type: "displayed-error", text: `You didn't enter anything to ${command}! Try entering: ${command} <something>` }] })
-
-                } else {
-                    let foundItem = user.inventory.find(param => {
-                        return param.item.itemName.includes(toExamine)
-                    })
-                    if (foundItem) {
-                        setChatHistory(prevState => {
-                            return [...prevState, { type: 'displayed-stat', text: `In your inventory you see ${foundItem.item.description}` }]
-                        })
-                    } else {
-                        setChatHistory(prevState => { return [...prevState, { type: "displayed-error", text: "There's nothing to discover by that name" }] })
-                    }
-                }
+                runExamine({ input, location, command, toExamine, user, setChatHistory });
 
             } else if (findIn(input, ["logout", "log out", "log off"])) {
                 takeTheseOffThat(["logout, log out", "log off"], input);
